update UserIdContext import paths and move context file
